scripts/agent/requirements.txt
anthropic
google-generativeai
playwright
python-dotenv
termcolor
